+ clean up more code warnings on int / float conversions, unused variables 	+ fixed an initialization pb in fluid that could potentially lead to a crash if a node is not recognized
git-svn-id: file:///fltk/svn/trunk@4882 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ test/arc.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: arc.cxx,v 1.21 2005/01/27 08:50:41 spitzak Exp $"
+// "$Id$"
 //
 // Arc drawing test program for the Fast Light Tool Kit (FLTK).
 //
@@ -49,16 +49,16 @@ class Drawing : public fltk::Widget {
     fltk::fillstrokepath(fltk::WHITE);
     // draw a hardware circle to see how well rotations match:
     fltk::setcolor(fltk::GRAY33);
-    fltk::addchord(Rectangle(20,20,args[2]+1,args[3]+1),args[4],args[5]);
+    fltk::addchord(Rectangle(20,20,(int)(args[2]+1),(int)(args[3]+1)),args[4],args[5]);
     fltk::fillstrokepath(fltk::WHITE);
     // now draw non-rotated hardware circle to check if it inscribes:
     fltk::pop_matrix();
     fltk::setcolor(fltk::GRAY40);
-    fltk::fillrect(fltk::Rectangle(10,270-args[3],args[2],args[3]));
+    fltk::fillrect(fltk::Rectangle(10,(int)(270-args[3]),(int)args[2],(int)args[3]));
     fltk::setcolor(fltk::GRAY90);
-    fltk::strokerect(fltk::Rectangle(10,270-args[3],args[2],args[3]));
+    fltk::strokerect(fltk::Rectangle(10,(int)(270-args[3]),(int) args[2],(int) args[3]));
     fltk::setcolor(fltk::GRAY10);
-    fltk::addchord(fltk::Rectangle(10,270-args[3],args[2],args[3]),args[4],args[5]);
+    fltk::addchord(fltk::Rectangle(10,(int) (270-args[3]),(int) args[2],(int)args[3]),args[4],args[5]);
     fltk::fillstrokepath(fltk::GRAY90);
     fltk::pop_clip();
   }
@@ -103,5 +103,5 @@ int main(int argc, char** argv) {
 
 
 //
-// End of "$Id: arc.cxx,v 1.21 2005/01/27 08:50:41 spitzak Exp $".
+// End of "$Id$".
 //
